Remove legacy layer Properties panel code
