+++ editor/src/messages/portfolio/document/properties_panel/properties_panel_message.rs
@@ -1,8 +1,4 @@
-use super::utility_types::TransformOp;
-use crate::messages::layout::utility_types::widget_prelude::*;
 use crate::messages::prelude::*;
-
-use document_legacy::LayerId;
 
 use serde::{Deserialize, Serialize};
 
@@ -11,15 +7,6 @@ use serde::{Deserialize, Serialize};
 #[derive(PartialEq, Clone, Debug, Serialize, Deserialize)]
 pub enum PropertiesPanelMessage {
 	// Messages
-	CheckSelectedWasDeleted { path: Vec<LayerId> },
-	CheckSelectedWasUpdated { path: Vec<LayerId> },
-	ClearSelection,
-	Deactivate,
-	Init,
-	ModifyPreserveAspect { preserve_aspect: bool },
-	ModifyTransform { value: f64, transform_op: TransformOp },
-	ResendActiveProperties,
-	SetActiveLayers { paths: Vec<Vec<LayerId>> },
-	SetPivot { new_position: PivotPosition },
-	UpdateSelectedDocumentProperties,
+	Clear,
+	Refresh,
 }
