+++ editor/src/messages/portfolio/document/properties_panel/utility_types.rs
@@ -1,8 +1,6 @@
 use document_legacy::document::Document as DocumentLegacy;
 use document_legacy::LayerId;
 
-use serde::{Deserialize, Serialize};
-
 use crate::{messages::prelude::NodeGraphMessageHandler, node_graph_executor::NodeGraphExecutor};
 
 pub struct PropertiesPanelMessageHandlerData<'a> {
@@ -11,15 +9,4 @@ pub struct PropertiesPanelMessageHandlerData<'a> {
 	pub selected_layers: &'a mut dyn Iterator<Item = &'a [LayerId]>,
 	pub node_graph_message_handler: &'a NodeGraphMessageHandler,
 	pub executor: &'a mut NodeGraphExecutor,
-}
-
-#[derive(PartialEq, Eq, Clone, Copy, Debug, Serialize, Deserialize, specta::Type)]
-pub enum TransformOp {
-	X,
-	Y,
-	ScaleX,
-	ScaleY,
-	Width,
-	Height,
-	Rotation,
 }
